pyproject.toml
[tool.poetry]
name="ticketer"
version="0.0.1"
description="A sample project to demostrate software testing for the Fundamentals of Software Design Course at Amirkabir University of Technology."
authors=["Ali Mehrabi <alimehrabi002@gmail.com>"]
readme="README.md"
packages=[{include="ticketer"}]

[tool.poetry.dependencies]
python=">=3.11,<3.12"
fastapi = "0.121.0"
uvicorn = "0.38.0"
sqlalchemy = "2.0.44"
alembic = "1.17.1"
pydantic = {version = "^2.12", extras = ["email"]}
pydantic-settings = "^2.11"
psycopg = {extras = ["binary"], version = "^3.2.12"}
python-jose = {version = "^3.5", extras = ["cryptography"]}
passlib = "^1.7"
bcrypt = "^5.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
python=">=3.11,<3.12"
ruff="^0.14.3"
ty="^0.0.1a25" 
pytest="^8.4.2"
pytest-mock="^3.15.1"
pytest-asyncio = "^1.2"
pytest-cov = "^7.0"
httpx = "^0.28"
factory-boy = "^3.3"

# Ruff configurations can go directly in pyproject.toml
[tool.ruff]
target-version = "py311"
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"